+++ spec/dummy/app/controllers/common_controller_actions.rb
@@ -8,7 +8,7 @@ module CommonControllerActions
     layout :layout_by_resource
     helper :all
     helper Koi::NavigationHelper
-    before_action :sign_in_as_admin! if Rails.env.development?
+    before_action :sign_in_as_admin!, if: -> { Rails.env.development? && !admin_signed_in? }
   end
 
   protected
@@ -24,7 +24,7 @@ module CommonControllerActions
 
   # FIXME: Hack to redirect back to admin after admin login
   def after_sign_in_path_for(resource_or_scope)
-    resource_or_scope.is_a?(Admin) ? koi_engine.root_path : super
+    resource_or_scope.is_a?(AdminUser) ? koi_engine.root_path : super
   end
 
   # FIXME: Hack to redirect back to admin after admin logout
@@ -33,6 +33,8 @@ module CommonControllerActions
   end
 
   def sign_in_as_admin!
-    sign_in(:admin, Admin.where(role: "Super").first) unless Admin.where(role: "Super").empty?
+    if (admin = AdminUser.find_by(role: "Super")).present?
+      sign_in :admin, admin
+    end
   end
 end